add: Firestore連携で使用するプロバイダクラス、プロバイダを作成

+++ lib/domain/repositories/implements/todo_repository_impl.dart
@@ -2,24 +2,32 @@
 
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:todo/domain/entities/todo_model.dart';
-import 'package:todo/domain/repositories/todo_repository_interface.dart';
+import 'package:todo/domain/repositories/interfaces/todo_repository_interface.dart';
 
 class FirebaseTodoRepository implements ITodoRepository {
   final FirebaseFirestore _firestore;
   final String _collection = 'todos';
+  final String? _userId;
 
-  FirebaseTodoRepository({FirebaseFirestore? firestore})
-      : _firestore = firestore ?? FirebaseFirestore.instance;
+  FirebaseTodoRepository(
+      {FirebaseFirestore? firestore, required String? userId})
+      : _firestore = firestore ?? FirebaseFirestore.instance,
+        _userId = userId;
 
-  // todosコレクションへの参照を取得
-  CollectionReference<Map<String, dynamic>> get _todosRef =>
-      _firestore.collection(_collection);
+  // コレクションの参照を取得（ユーザIDごとに分ける）
+  CollectionReference<Map<String, dynamic>> get _todosRef {
+    if (_userId == null) {
+      throw Exception('ユーザーがログインしていません');
+    }
+    return _firestore.collection('users/$_userId/$_collection');
+  }
 
-  // todosコレクションにアクセス、createdDateフィールドで降順（新しい順）に並び替え
-  // リアルタイム更新のストリームを取得
-  // 取得したデータを加工（map）
   @override
   Stream<List<TodoModel>> getAllTodos() {
+    if (_userId == null) {
+      return Stream.value([]);
+    }
+
     return _todosRef
         .orderBy('createdDate', descending: true)
         .snapshots()
@@ -32,6 +40,10 @@ class FirebaseTodoRepository implements ITodoRepository {
 
   @override
   Stream<List<TodoModel>> getCompletedTodo() {
+    if (_userId == null) {
+      return Stream.value([]);
+    }
+
     return _todosRef
         .where('isCompleted', isEqualTo: true)
         .orderBy('createdDate', descending: true)
@@ -45,6 +57,10 @@ class FirebaseTodoRepository implements ITodoRepository {
 
   @override
   Stream<List<TodoModel>> getInCompletedTodo() {
+    if (_userId == null) {
+      return Stream.value([]);
+    }
+
     return _todosRef
         .where('isCompleted', isEqualTo: false)
         .orderBy('createdDate', descending: true)
@@ -58,22 +74,38 @@ class FirebaseTodoRepository implements ITodoRepository {
 
   @override
   Future<String> addTodo(TodoModel todoModel) async {
+    if (_userId == null) {
+      throw Exception('ユーザーがログインしていません');
+    }
+
     final docRef = await _todosRef.add(TodoModel.toFirestore(todoModel));
     return docRef.id;
   }
 
   @override
   Future<void> updateTodo(TodoModel todoModel) async {
+    if (_userId == null) {
+      throw Exception('ユーザーがログインしていません');
+    }
+
     await _todosRef.doc(todoModel.id).update(TodoModel.toFirestore(todoModel));
   }
 
   @override
   Future<void> deleteTodo(String id) async {
+    if (_userId == null) {
+      throw Exception('ユーザーがログインしていません');
+    }
+
     await _todosRef.doc(id).delete();
   }
 
   @override
   Future<void> setTodoCompleted(TodoModel todoModel, bool isCompelted) async {
+    if (_userId == null) {
+      throw Exception('ユーザーがログインしていません');
+    }
+
     final updatedTodo = todoModel.copyWith(isCompleted: isCompelted);
     await _todosRef
         .doc(todoModel.id)
